refactor(server): rename the join-cardinality util to hasOnlyToOneJoins
